refactor: extract StateUpdateBuilder into its own file
Move StateUpdateBuilder from consume_ticks.dart into
state_update_builder.dart. Move ticksPer1Hp constant to tick.dart.
Replace direct _state access in agility processing with the existing
updateActivity() public method.

This removes a circular dependency barrier, enabling data-layer classes
to accept StateUpdateBuilder as a parameter in the future.
